requirements.txt
python-telegram-bot==20.6
Flask==2.2.5
playwright==1.42.0
